Audit fix (#22)
* fix(audit): fixes https-proxy-agent vulnerability

This is a hack, but will solve the problem for a while. For more context
on what was done, refer to:
https://github.com/sass/node-sass/issues/2625#issuecomment-482224111

The process I used is similar.

* fix(audit): applies temporary package-lock fix

Used the technique described in https://github.com/sass/node-sass/issues/2625#issuecomment-482224111
This is supposed to be a temporary fix to make npm audit pass - remove once @semantic-version/npm is patched.